Dose response curve helper function

This function peforms the four-parameter dose response curve fit. It is the helper function for the fit in the fit_drc_4p function.

drc_4p(data, response, dose, log_logarithmic = TRUE, pb = NULL)

Arguments

data

a data frame that contains at least the dose and response column the model should be fitted to.

response

a numeric column that contains the response values.

dose

a numeric column that contains the dose values.

log_logarithmic

a logical value indicating if a logarithmic or log-logarithmic model is fitted. If response values form a symmetric curve for non-log transformed dose values, a logarithmic model instead of a log-logarithmic model should be used. Usually biological dose response data has a log-logarithmic distribution, which is the reason this is the default. Log-logarithmic models are symmetric if dose values are log transformed.

pb

progress bar object. This is only necessary if the function is used in an iteration.

Value

An object of class drc. If no fit was performed a character vector with content "no_fit".