Fetch AlphaFold aligned error

Fetches the aligned error for AlphaFold predictions for provided proteins. The aligned error is useful for assessing inter-domain accuracy. In detail it represents the expected position error at residue x (scored residue), when the predicted and true structures are aligned on residue y (aligned residue).

fetch_alphafold_aligned_error(
  uniprot_ids = NULL,
  error_cutoff = 20,
  timeout = 3600,
  return_data_frame = FALSE,
  show_progress = TRUE
)

Arguments

uniprot_ids

a character vector of UniProt identifiers for which predictions should be fetched.

error_cutoff

a numeric value specifying the maximum position error (in Angstroms) that should be retained. setting this value to a low number reduces the size of the retrieved data. Default is 20.

timeout

a numeric value specifying the time in seconds until the download of an organism archive times out. The default is 3600 seconds.

return_data_frame

a logical value; if TRUE a data frame instead of a list is returned. It is recommended to only use this if information for few proteins is retrieved. Default is FALSE.

show_progress

a logical value; if TRUE a progress bar will be shown. Default is TRUE.

Value

A list that contains aligned errors for AlphaFold predictions. If return_data_frame is TRUE, a data frame with this information is returned instead. The data frame contains the following columns:

scored_residue: The error for this position is calculated based on the alignment to the aligned residue.

aligned_residue: The residue that is aligned for the calculation of the error of the scored residue

error: The predicted aligned error computed by alpha fold.

accession: The UniProt protein identifier.

Examples

# \donttest{
aligned_error <- fetch_alphafold_aligned_error(uniprot_ids = c("F4HVG8", "O15552"),
                                               error_cutoff = 5,
                                               return_data_frame = TRUE)

head(aligned_error, n = 10)
#> # A tibble: 10 × 4
#>    aligned_residue scored_residue error accession
#>              <int>          <int> <dbl> <chr>    
#>  1               1              1     0 F4HVG8   
#>  2               1              2     1 F4HVG8   
#>  3               1              3     3 F4HVG8   
#>  4               1              4     4 F4HVG8   
#>  5               1              5     4 F4HVG8   
#>  6               1              6     4 F4HVG8   
#>  7               2              1     1 F4HVG8   
#>  8               2              2     0 F4HVG8   
#>  9               2              3     1 F4HVG8   
#> 10               2              4     2 F4HVG8   
# }