Fetch ChEBI database information

Fetches information from the ChEBI database.

fetch_chebi(relation = FALSE, stars = c(3), timeout = 60)

Arguments

relation

a logical value that indicates if ChEBI Ontology data will be returned instead the main compound data. This data can be used to check the relations of ChEBI ID's to each other. Default is FALSE.

stars

a numeric vector indicating the "star" level (confidence) for which entries should be retrieved (Possible levels are 1, 2 and 3). Default is c(3) retrieving only "3-star" entries, which are manually annotated by the ChEBI curator team.

timeout

a numeric value specifying the time in seconds until the download of an organism archive times out. The default is 60 seconds.

Value

A data frame that contains information about each molecule in the ChEBI database.

Examples

# \donttest{
chebi <- fetch_chebi()

head(chebi)
#>   id chebi_accession parent_id                         definition star
#> 1  1         CHEBI:1     18357 The R-enantiomer of noradrenaline.    3
#> 2  1         CHEBI:1     18357 The R-enantiomer of noradrenaline.    3
#> 3  1         CHEBI:1     18357 The R-enantiomer of noradrenaline.    3
#> 4  1         CHEBI:1     18357 The R-enantiomer of noradrenaline.    3
#> 5  1         CHEBI:1     18357 The R-enantiomer of noradrenaline.    3
#> 6  1         CHEBI:1     18357 The R-enantiomer of noradrenaline.    3
#>            name type_name            type_accession accession_number  formula
#> 1 Noradrenaline   SYNONYM   KEGG COMPOUND accession           C00547 C8H11NO3
#> 2 Noradrenaline   SYNONYM       CAS Registry Number          51-41-2 C8H11NO3
#> 3 Noradrenaline   SYNONYM        PDBeChem accession              LNR C8H11NO3
#> 4 Noradrenaline   SYNONYM Beilstein Registry Number          4231961 C8H11NO3
#> 5 Noradrenaline   SYNONYM Beilstein Registry Number          2804840 C8H11NO3
#> 6 Noradrenaline   SYNONYM       KEGG DRUG accession           D00076 C8H11NO3
#>        mass charge monoisotopic_mass
#> 1 169.17780      0         169.07389
#> 2 169.17780      0         169.07389
#> 3 169.17780      0         169.07389
#> 4 169.17780      0         169.07389
#> 5 169.17780      0         169.07389
#> 6 169.17780      0         169.07389
# }