Fetch evidence & conclusion ontology

Fetches all evidence & conclusion ontology (ECO) information from the QuickGO EBI database. The ECO project is maintained through a public GitHub repository.

fetch_eco(
  return_relation = FALSE,
  return_history = FALSE,
  show_progress = TRUE
)

Arguments

return_relation

a logical value that indicates if relational information should be returned instead the main descriptive information. This data can be used to check the relations of ECO terms to each other. Default is FALSE.

return_history

a logical value that indicates if the entry history of an ECO term should be returned instead the main descriptive information. Default is FALSE.

show_progress

a logical value that indicates if a progress bar will be shown. Default is TRUE.

Value

A data frame that contains descriptive information about each ECO term in the EBI database. If either return_relation or return_history is set to TRUE, the respective information is returned instead of the usual output.

Details

According to the GitHub repository ECO is defined as follows:

"The Evidence & Conclusion Ontology (ECO) describes types of scientific evidence within the biological research domain that arise from laboratory experiments, computational methods, literature curation, or other means. Researchers use evidence to support conclusions that arise out of scientific research. Documenting evidence during scientific research is essential, because evidence gives us a sense of why we believe what we think we know. Conclusions are asserted as statements about things that are believed to be true, for example that a protein has a particular function (i.e. a protein functional annotation) or that a disease is associated with a particular gene variant (i.e. a phenotype-gene association). A systematic and structured (i.e. ontological) classification of evidence allows us to store, retreive, share, and compare data associated with that evidence using computers, which are essential to navigating the ever-growing (in size and complexity) corpus of scientific information."

More information can be found in their publication.

Examples

# \donttest{
eco <- fetch_eco()

head(eco)
#> # A tibble: 6 × 10
#>   id          is_obs…¹ main_…² defin…³ comment name  type  db_code db_id secon…⁴
#>   <chr>       <lgl>    <chr>   <chr>   <chr>   <chr> <chr> <chr>   <chr> <chr>  
#> 1 ECO:0000005 FALSE    enzyma… A type… NA      enzy… rela… MI      0415  ECO:00…
#> 2 ECO:0000005 FALSE    enzyma… A type… NA      enzy… exact MI      0415  ECO:00…
#> 3 ECO:0000226 FALSE    chroma… A type… Chroma… ChIP… exact MI      0402  NA     
#> 4 ECO:0000293 FALSE    system… A type… SELEX … SELE… exact MI      0657  NA     
#> 5 ECO:0000293 FALSE    system… A type… SELEX … in v… exact MI      0657  NA     
#> 6 ECO:0000293 FALSE    system… A type… SELEX … in v… rela… MI      0657  NA     
#> # … with abbreviated variable names ¹​is_obsolete, ²​main_name, ³​definition,
#> #   ⁴​secondary_ids
# }