Fetch protein disorder and mobility information from MobiDB

Fetches information about disordered and flexible protein regions from MobiDB.

fetch_mobidb(
  uniprot_ids = NULL,
  organism_id = NULL,
  show_progress = TRUE,
  timeout = 60,
  max_tries = 2
)

Arguments

uniprot_ids

optional, a character vector of UniProt identifiers for which information should be fetched. This argument is mutually exclusive to the organism_id argument.

organism_id

optional, a character value providing the NCBI taxonomy identifier of an organism (TaxId) of an organism for which all available information should be retreived. This argument is mutually exclusive to the uniprot_ids argument.

show_progress

a logical value; if TRUE a progress bar will be shown. Default is TRUE.

timeout

a numeric value specifying the time in seconds until the download of an organism archive times out. The default is 60 seconds.

max_tries

a numeric value that specifies the number of times the function tries to download the data in case an error occurs. The default is 2.

Value

A data frame that contains start and end positions for disordered and flexible protein regions. The feature column contains information on the source of this annotation. More information on the source can be found here.

Examples

# \donttest{
fetch_mobidb(
  uniprot_ids = c("P0A799", "P62707")
)
#> Consider increasing the "timeout" or "max_tries" argument. 
#> The following IDs have not been retrieved correctly.
#> # A tibble: 1 × 2
#>   id          error                                                             
#>   <chr>       <chr>                                                             
#> 1 "IDs:  to " "Timeout was reached [mobidb.bio.unipd.it]:\nConnection timed out…
#> No valid information was retrieved!
# }