Fetch information from the QuickGO API

Fetches gene ontology (GO) annotations, terms or slims from the QuickGO EBI database. Annotations can be retrieved for specific UniProt IDs or NCBI taxonomy identifiers. When terms are retrieved, a complete list of all GO terms is returned. For the generation of a slim dataset you can provide GO IDs that should be considered. A slim dataset is a subset GO dataset that considers all child terms of the supplied IDs.

fetch_quickgo(
  type = "annotations",
  id_annotations = NULL,
  taxon_id_annotations = NULL,
  ontology_annotations = "all",
  go_id_slims = NULL,
  relations_slims = c("is_a", "part_of", "regulates", "occurs_in"),
  show_progress = TRUE
)

Arguments

type

a character value that indicates if gene ontology terms, annotations or slims should be retrieved. The possible values therefore include "annotations", "terms" and "slims". If annotations are retrieved, the maximum number of results is 2,000,000.

id_annotations

an optional character vector that specifies UniProt IDs for which GO annotations should be retrieved. This argument should only be provided if annotations are retrieved.

taxon_id_annotations

an optional character value that specifies the NCBI taxonomy identifier (TaxId) for an organism for which GO annotations should be retrieved. This argument should only be provided if annotations are retrieved.

ontology_annotations

an optional character value that specifies the ontology that should be retrieved. This can either have the values "all", "molecular_function", "biological_process" or "cellular_component". This argument should only be provided if annotations are retrieved.

go_id_slims

an optional character vector that specifies gene ontology IDs (e.g. GO:0046872) for which a slim go set should be generated. This argument should only be provided if slims are retrieved.

relations_slims

an optional character vector that specifies the relations of GO IDs that should be considered for the generation of the slim dataset. This argument should only be provided if slims are retrieved.

show_progress

a logical value that indicates if a progress bar will be shown. Default is TRUE.

Value

A data frame that contains descriptive information about gene ontology annotations, terms or slims depending on what the input "type" was.

Examples

# \donttest{
# Annotations
annotations <- fetch_quickgo(
  type = "annotations",
  id = c("P63328", "Q4FFP4"),
  ontology = "molecular_function"
)
#> Retrieving GO annotations ... 
#> DONE(2.64s)

head(annotations)
#> # A tibble: 6 × 15
#>   gene_p…¹ gene_…² go_name symbol quali…³ go_term go_as…⁴ eco_id go_ev…⁵ refer…⁶
#>   <chr>    <chr>   <chr>   <chr>  <chr>   <chr>   <chr>   <chr>  <chr>   <chr>  
#> 1 UniProt… P63328  calmod… Ppp3ca enables GO:000… molecu… ECO:0… IBA     PMID:2…
#> 2 UniProt… P63328  calmod… Ppp3ca enables GO:003… molecu… ECO:0… IBA     PMID:2…
#> 3 UniProt… P63328  protei… Ppp3ca enables GO:000… molecu… ECO:0… IPI     PMID:3…
#> 4 UniProt… P63328  protei… Ppp3ca enables GO:000… molecu… ECO:0… IPI     PMID:1…
#> 5 UniProt… P63328  calmod… Ppp3ca enables GO:003… molecu… ECO:0… ISS     GO_REF…
#> 6 UniProt… P63328  hydrol… Ppp3ca enables GO:001… molecu… ECO:0… IEA     GO_REF…
#> # … with 5 more variables: with_from <chr>, taxon_id <dbl>, assigned_by <chr>,
#> #   annotation_extension <lgl>, date <dbl>, and abbreviated variable names
#> #   ¹​gene_product_db, ²​gene_product_id, ³​qualifier, ⁴​go_aspect,
#> #   ⁵​go_evidence_code, ⁶​reference

# Terms
terms <- fetch_quickgo(type = "terms")

head(terms)
#> # A tibble: 6 × 13
#>   main_id  is_ob…¹ main_…² defin…³ ontol…⁴ usage child…⁵ child…⁶ chebi…⁷ relat…⁸
#>   <chr>    <lgl>   <chr>   <chr>   <chr>   <chr> <chr>   <chr>   <chr>   <chr>  
#> 1 GO:0006… FALSE   creati… The ch… biolog… Unre… GO:000… is_a;i… CHEBI:… creati…
#> 2 GO:0006… FALSE   creati… The ch… biolog… Unre… NA      NA      CHEBI:… creati…
#> 3 GO:0006… FALSE   protei… The di… biolog… Unre… GO:004… regula… NA      NA     
#> 4 GO:0006… FALSE   protei… The pr… biolog… Unre… GO:000… is_a;i… NA      NA     
#> 5 GO:0006… FALSE   riboso… The di… biolog… Unre… NA      NA      NA      NA     
#> 6 GO:0006… FALSE   phosph… The ch… biolog… Unre… GO:004… is_a;i… CHEBI:… N(omeg…
#> # … with 3 more variables: database <chr>, relations_url <chr>,
#> #   relations_relation <chr>, and abbreviated variable names ¹​is_obsolete,
#> #   ²​main_name, ³​definition, ⁴​ontology, ⁵​child_id, ⁶​children_relation,
#> #   ⁷​chebi_id, ⁸​relations_term

# Slims
slims <- fetch_quickgo(
  type = "slims",
  go_id_slims = c("GO:0046872", "GO:0051540")
)

head(slims)
#> # A tibble: 6 × 2
#>   slims_from_id slims_to_ids
#>   <chr>         <chr>       
#> 1 GO:0046914    GO:0046872  
#> 2 GO:0032791    GO:0046872  
#> 3 GO:1904434    GO:0046872  
#> 4 GO:1904433    GO:0046872  
#> 5 GO:1904432    GO:0046872  
#> 6 GO:0030151    GO:0046872  
# }