Fetch protein data from UniProt

Fetches protein metadata from UniProt.

fetch_uniprot(
  uniprot_ids,
  columns = c("protein names", "length", "sequence", "genes", "database(GeneID)",
    "database(String)", "go(molecular function)", "go(biological process)",
    "go(cellular compartment)", "interactor", "feature(ACTIVE SITE)",
    "feature(BINDING SITE)", "feature(METAL BINDING)", "chebi(Cofactor)",
    "chebi(Catalytic activity)", "database(PDB)"),
  batchsize = 200,
  show_progress = TRUE
)

Arguments

uniprot_ids
A character vector of UniProt accession numbers

columns
Metadata columns that should be imported from UniProt (all possible columns can be found here.)

batchsize
Size of batch of proteins for a single query

show_progress
Logical, if true, a progress bar will be shown

Value

A data frame that contains all protein metadata specified in columns for the proteins provided. If an invalid ID was provided that contains a valid UniProt ID, the valid portion of the ID is fetched and the invalid input ID is saved in a column called input_id.

Examples

# \donttest{
fetch_uniprot(c("P36578", "O43324", "Q00796"))

#>       id
#> 1 P36578
#> 2 O43324
#> 3 Q00796
#>                                                                                                                                                                                                                                    protein_names
#> 1                                                                                                                                                      60S ribosomal protein L4 (60S ribosomal protein L1) (Large ribosomal subunit protein uL4)
#> 2                                         Eukaryotic translation elongation factor 1 epsilon-1 (Aminoacyl tRNA synthetase complex-interacting multifunctional protein 3) (Elongation factor p18) (Multisynthase complex auxiliary component p18)
#> 3 Sorbitol dehydrogenase (SDH) (EC 1.1.1.-) ((R,R)-butanediol dehydrogenase) (EC 1.1.1.4) (L-iditol 2-dehydrogenase) (EC 1.1.1.14) (Polyol dehydrogenase) (Ribitol dehydrogenase) (RDH) (EC 1.1.1.56) (Xylitol dehydrogenase) (XDH) (EC 1.1.1.9)
#>   length
#> 1    427
#> 2    174
#> 3    357
#>                                                                                                                                                                                                                                                                                                                                                                                                                                      sequence
#> 1 MACARPLISVYSEKGESSGKNVTLPAVFKAPIRPDIVNFVHTNLRKNNRQPYAVSELAGHQTSAESWGTGRAVARIPRVRGGGTHRSGQGAFGNMCRGGRMFAPTKTWRRWHRRVNTTQKRYAICSALAASALPALVMSKGHRIEEVPELPLVVEDKVEGYKKTKEAVLLLKKLKAWNDIKKVYASQRMRAGKGKMRNRRRIQRRGPCIIYNEDNGIIKAFRNIPGITLLNVSKLNILKLAPGGHVGRFCIWTESAFRKLDELYGTWRKAASLKSNYNLPMHKMINTDLSRILKSPEIQRALRAPRKKIHRRVLKKNPLKNLRIMLKLNPYAKTMRRNTILRQARNHKLRVDKAAAAAAALQAKSDEKAAVAGKKPVVGKKGKKAAVGVKKQKKPLVGKKAAATKKPAPEKKPAEKKPTTEEKKPAA
#> 2                                                                                                                                                                                                                                                              MAAAAELSLLEKSLGLSKGNKYSAQGERQIPVLQTNNGPSLTGLTTIAAHLVKQANKEYLLGSTAEEKAIVQQWLEYRVTQVDGHSSKNDIHTLLKDLNSYLEDKVYLTGYNFTLADILLYYGLHRFIVDLTVQEKEKYLNVSRWFCHIQHYPGIRQHLSSVVFIKNRLYTNSH
#> 3                                                                       MAAAAKPNNLSLVVHGPGDLRLENYPIPEPGPNEVLLRMHSVGICGSDVHYWEYGRIGNFIVKKPMVLGHEASGTVEKVGSSVKHLKPGDRVAIEPGAPRENDEFCKMGRYNLSPSIFFCATPPDDGNLCRFYKHNAAFCYKLPDNVTFEEGALIEPLSVGIHACRRGGVTLGHKVLVCGAGPIGMVTLLVAKAMGAAQVVVTDLSATRLSKAKEIGADLVLQISKESPQEIARKVEGQLGCKPEVTIECTGAEASIQAGIYATRSGGNLVLVGLGSEMTTVPLLHAAIREVDIKGVFRYCNTWPVAISMLASKSVNVKPLVTHRFPLEKALEAFETFKKGLGLKIMLKCDPSDQNP
#>              genes database_gene_id       database_string
#> 1        RPL4 RPL1            6124; 9606.ENSP00000311430;
#> 2 EEF1E1 AIMP3 P18            9521; 9606.ENSP00000369038;
#> 3             SORD            6652; 9606.ENSP00000267814;
#>                                                                                                                                                                                                                                                                                                                 go_molecular_function
#> 1                                                                                                                                                                                                                                                           RNA binding [GO:0003723]; structural constituent of ribosome [GO:0003735]
#> 2                                                                                                                                                                                                                                                                                                                                    
#> 3 (R,R)-butanediol dehydrogenase activity [GO:0000721]; carbohydrate binding [GO:0030246]; D-xylulose reductase activity [GO:0046526]; identical protein binding [GO:0042802]; L-iditol 2-dehydrogenase activity [GO:0003939]; NAD binding [GO:0051287]; ribitol 2-dehydrogenase activity [GO:0050255]; zinc ion binding [GO:0008270]
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         go_biological_process
#> 1                                                                                                                                                                                                                                                                                               nuclear-transcribed mRNA catabolic process, nonsense-mediated decay [GO:0000184]; SRP-dependent cotranslational protein targeting to membrane [GO:0006614]; translation [GO:0006412]; translational initiation [GO:0006413]; viral transcription [GO:0019083]
#> 2                                                                                                                                                                                                             negative regulation of cell population proliferation [GO:0008285]; positive regulation of apoptotic process [GO:0043065]; positive regulation of cellular senescence [GO:2000774]; positive regulation of DNA damage response, signal transduction by p53 class mediator [GO:0043517]; tRNA aminoacylation for protein translation [GO:0006418]
#> 3 flagellated sperm motility [GO:0030317]; fructose biosynthetic process [GO:0046370]; glucose metabolic process [GO:0006006]; glucuronate catabolic process to xylulose 5-phosphate [GO:0019640]; L-xylitol catabolic process [GO:0051160]; L-xylitol metabolic process [GO:0051164]; response to cadmium ion [GO:0046686]; response to copper ion [GO:0046688]; response to drug [GO:0042493]; response to hormone [GO:0009725]; response to nutrient levels [GO:0031667]; response to osmotic stress [GO:0006970]; sorbitol catabolic process [GO:0006062]
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                go_cellular_compartment
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                       cytoplasm [GO:0005737]; cytosol [GO:0005829]; cytosolic large ribosomal subunit [GO:0022625]; extracellular exosome [GO:0070062]; focal adhesion [GO:0005925]; membrane [GO:0016020]; nucleolus [GO:0005730]; nucleus [GO:0005634]; ribonucleoprotein complex [GO:1990904]; rough endoplasmic reticulum [GO:0005791]; RNA binding [GO:0003723]; structural constituent of ribosome [GO:0003735]; nuclear-transcribed mRNA catabolic process, nonsense-mediated decay [GO:0000184]; SRP-dependent cotranslational protein targeting to membrane [GO:0006614]; translation [GO:0006412]; translational initiation [GO:0006413]; viral transcription [GO:0019083]
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    aminoacyl-tRNA synthetase multienzyme complex [GO:0017101]; cytoplasm [GO:0005737]; cytosol [GO:0005829]; nucleolus [GO:0005730]; nucleoplasm [GO:0005654]; nucleus [GO:0005634]; negative regulation of cell population proliferation [GO:0008285]; positive regulation of apoptotic process [GO:0043065]; positive regulation of cellular senescence [GO:2000774]; positive regulation of DNA damage response, signal transduction by p53 class mediator [GO:0043517]; tRNA aminoacylation for protein translation [GO:0006418]
#> 3 cytosol [GO:0005829]; extracellular exosome [GO:0070062]; extracellular space [GO:0005615]; membrane [GO:0016020]; mitochondrial membrane [GO:0031966]; motile cilium [GO:0031514]; (R,R)-butanediol dehydrogenase activity [GO:0000721]; carbohydrate binding [GO:0030246]; D-xylulose reductase activity [GO:0046526]; identical protein binding [GO:0042802]; L-iditol 2-dehydrogenase activity [GO:0003939]; NAD binding [GO:0051287]; ribitol 2-dehydrogenase activity [GO:0050255]; zinc ion binding [GO:0008270]; flagellated sperm motility [GO:0030317]; fructose biosynthetic process [GO:0046370]; glucose metabolic process [GO:0006006]; glucuronate catabolic process to xylulose 5-phosphate [GO:0019640]; L-xylitol catabolic process [GO:0051160]; L-xylitol metabolic process [GO:0051164]; response to cadmium ion [GO:0046686]; response to copper ion [GO:0046688]; response to drug [GO:0042493]; response to hormone [GO:0009725]; response to nutrient levels [GO:0031667]; response to osmotic stress [GO:0006970]; sorbitol catabolic process [GO:0006062]
#>                                       interactor feature_active_site
#> 1 P42858; Q99558; P50914; Q02543; P46779; P60866                  NA
#> 2                         Q13137; Q9BTE0; P04591                  NA
#> 3                                         Itself                  NA
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   feature_binding_site
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#> 3 BINDING 51;  /note="Substrate";  /evidence="ECO:0000250|UniProtKB:P07846"; BINDING 156;  /note="Substrate";  /evidence="ECO:0000250|UniProtKB:P07846"; BINDING 184;  /note="NAD; via amide nitrogen";  /evidence="ECO:0000244|PDB:1PL6, ECO:0000244|PDB:1PL8, ECO:0000269|PubMed:12962626"; BINDING 204;  /note="NAD";  /evidence="ECO:0000244|PDB:1PL6, ECO:0000244|PDB:1PL8, ECO:0000269|PubMed:12962626"; BINDING 209;  /note="NAD";  /evidence="ECO:0000244|PDB:1PL6, ECO:0000244|PDB:1PL8, ECO:0000269|PubMed:12962626"; BINDING 299;  /note="Substrate";  /evidence="ECO:0000250|UniProtKB:P07846"; BINDING 300;  /note="Substrate";  /evidence="ECO:0000250|UniProtKB:P07846"
#>                                                                                                                                                                                                                                                                                                                                                          feature_metal_binding
#> 1                                                                                                                                                                                                                                                                                                                                                                             
#> 2                                                                                                                                                                                                                                                                                                                                                                             
#> 3 METAL 45;  /note="Zinc; catalytic";  /evidence="ECO:0000244|PDB:1PL7, ECO:0000244|PDB:1PL8, ECO:0000269|PubMed:12962626"; METAL 70;  /note="Zinc; catalytic";  /evidence="ECO:0000244|PDB:1PL7, ECO:0000244|PDB:1PL8, ECO:0000269|PubMed:12962626"; METAL 71;  /note="Zinc; catalytic";  /evidence="ECO:0000244|PDB:1PL7, ECO:0000244|PDB:1PL8, ECO:0000269|PubMed:12962626"
#>         chebi_cofactor
#> 1                     
#> 2                     
#> 3 Zn(2+) [CHEBI:29105]
#>                                                                                                                                                                                                                                                                                                                                                                          chebi_catalytic_activity
#> 1                                                                                                                                                                                                                                                                                                                                                                                                
#> 2                                                                                                                                                                                                                                                                                                                                                                                                
#> 3 D-sorbitol [CHEBI:17924]; NAD(+) [CHEBI:57540]; (R)-acetoin [CHEBI:15686]; L-erythrulose [CHEBI:27913]; L-threitol [CHEBI:42090]; H(+) [CHEBI:15378]; D-xylulose [CHEBI:17140]; keto-L-sorbose [CHEBI:13172]; D-ribulose [CHEBI:17173]; (R,R)-butane-2,3-diol [CHEBI:16982]; D-fructose [CHEBI:37721]; NADH [CHEBI:57945]; L-iditol [CHEBI:18202]; ribitol [CHEBI:15963]; xylitol [CHEBI:17151]
#>                                                                                                                                                                      database_pdb
#> 1 4UG0;4V6X;5A8L;5AJ0;5LKS;5T2C;6EK0;6IP5;6IP6;6IP8;6LQM;6LSR;6LSS;6LU8;6OLE;6OLF;6OLG;6OLI;6OLZ;6OM0;6OM7;6QZP;6W6L;6XA1;6Y0G;6Y2L;6Y57;6Y6X;6Z6L;6Z6M;6Z6N;6ZM7;6ZME;6ZMI;6ZMO;
#> 2                                                                                                                                                  2UZ8;4BL7;4BVX;4BVY;5BMU;5Y6L;
#> 3                                                                                                                                                                 1PL6;1PL7;1PL8;

# }