Find all ChEBI sub IDs of an ID

For a given ChEBI ID, find all ChEBI sub IDs (incoming IDs) and their sub IDs. The type of relationship can be selected too. This is a helper function for other functions.

find_all_subs(data, id, type = "is_a")

Arguments

data
A data frame that contains information on ChEBI IDs (id), their sub IDs (incoming) and their relationship (type). This data frame can be obtained by calling fetch_chebi(relation = TRUE).

id
A character vector of ChEBI IDs for which sub IDs should be retreived.

type
A character vector containing the type of relationship that should be considered for retreival. It is possible to use "all" relationships. The default type is "is_a". A list of possible relationships can be found here.

Value

A list of character vector containing the provided ID and all of its sub IDs. It contains one element per input ID.

Examples

if (FALSE) {
find_all_subs(
data,
id = c("25213", "30408"),
type = "is_a"
)
}