Perform KEGG pathway enrichment analysis

Analyses enrichment of KEGG pathways associated with proteins in the fraction of significant proteins compared to all detected proteins. A Fisher's exact test is performed to test significance of enrichment.

kegg_enrichment(
  data,
  protein_id,
  is_significant,
  pathway_id = pathway_id,
  pathway_name = pathway_name,
  plot = TRUE,
  plot_cutoff = "adj_pval top10"
)

Arguments

data
A data frame that contains at least the input variables.

protein_id
The name of the column containing the protein accession numbers.

is_significant
The name of the column containing a logical indicating if the corresponding protein has a significantly changing peptide. The input data frame may contain peptide level information with significance information. The function is able to extract protein level information from this.

pathway_id
The name of the column containing KEGG pathway identifiers. These can be obtained from KEGG using fetch_kegg.

pathway_name
The name of the column containing KEGG pathway names. These can be obtained from KEGG using fetch_kegg.

plot
A logical indicating whether the result should be plotted or returned as a table.

plot_cutoff
A character vector indicating if the plot should contain the top 10 most significant proteins (p-value or adjusted p-value), or if a significance cutoff should be used to determine the number of GO terms in the plot. This information should be provided with the type first followed by the threshold separated by a space. Example are plot_cutoff = "adj_pval top10", plot_cutoff = "pval 0.05" or plot_cutoff = "adj_pval 0.01". The threshold can be chosen freely.

Value

A bar plot displaying negative log10 adjusted p-values for the top 10 enriched pathways. Bars are coloured according to the direction of the enrichment. If plot = FALSE, a data frame is returned.

Examples

if (FALSE) {
kegg_enrichment(
data,
protein_id = pg_protein_accessions,
is_significant = significant,
pathway_id = pathway_id,
pathway_name = pathway_name
)
}