Assign peptide type

Based on preceding and C-terminal amino acid, the peptide type of a given peptide is assigned. Peptides with preceeding and C-terminal lysine or arginine are considered fully-tryptic. If a peptide is located at the N- or C-terminus of a protein and fulfills the criterium to be fully-tryptic otherwise, it is also considered as fully-tryptic. Peptides that only fulfill the criterium on one terminus are semi-tryptic peptides. Lastly, peptides that are not fulfilling the criteria for both termini are non-tryptic peptides.

peptide_type(
  data,
  aa_before = aa_before,
  last_aa = last_aa,
  aa_after = aa_after
)

Arguments

data
A data frame containing at least information about the preceding and C-terminal amino acids of peptides.

aa_before
The name of the column containing the preceding amino acid as one letter code.

last_aa
The name of the column containing the C-terminal amino acid as one letter code.

aa_after
The name of the column containing the following amino acid as one letter code.

Value

A data frame that contains the input data and an additional column with the peptide type information.

Examples

data <- data.frame(
  aa_before = c("K", "S", "T"),
  last_aa = c("R", "K", "Y"),
  aa_after = c("T", "R", "T")
)

peptide_type(data, aa_before, last_aa, aa_after)

#>   aa_before last_aa aa_after      pep_type
#> 1         K       R        T fully-tryptic
#> 2         S       K        R  semi-tryptic
#> 3         T       Y        T   non-tryptic