Plotting of four-parameter dose response curves

Function for plotting four-parameter dose response curves for each group (precursor, peptide or protein), based on output from fit_drc_4p function.

plot_drc_4p(
  data,
  grouping,
  response,
  dose,
  targets,
  unit = "uM",
  y_axis_name = "Response",
  facet = TRUE,
  scales = "free",
  x_axis_scale_log10 = TRUE,
  export = FALSE,
  export_name = "dose-response_curves"
)

Arguments

data
A data frame that is obtained by calling the fit_drc_4p function.

grouping
The name of the column containing precursor, peptide or protein identifiers.

response
The name of the column containing response values, eg. log2 transformed intensities.

dose
The name of the column containing dose values, eg. the treatment concentrations.

targets
A character vector that specifies the names of the precursors, peptides or proteins (depending on grouping) that should be plotted. This can also be "all" if plots for all curve fits should be created.

unit
A character vector specifying the unit of the concentration.

y_axis_name
A character vector specifying the name of the y-axis of the plot.

facet
A logical indicating if plots should be summarised into facets of 20 plots. This is recommended for many plots.

scales
A character vector that specifies if the scales in faceted plots (if more than one target was provided) should be "free" or "fixed".

x_axis_scale_log10
A logical indicating if the x-axis scale should be log10 transformed.

export
A logical indicating if plots should be exported as PDF. The output directory will be the current working directory. The name of the file can be chosen using the export_name argument. If only one target is selected and export = TRUE, the plot is exported and in addition returned in R.

export_name
A character vector providing the name of the exported file if export = TRUE.

Value

If targets = "all" a list containing plots for every unique identifier in the grouping variable is created. Otherwise a plot for the specified targets is created with maximally 20 facets.

Examples

if (FALSE) {
plot_drc_4p(
  data,
  grouping = eg_precursor_id,
  response = intensity,
  dose = concentration,
  targets = c("ABCDEFK")
)
}