Perform normalization of LiP peptide data by protein abundance from tryptic controls

LiP peptide intensities are normalised by protein abundances from tryptic controls to correct for protein fold changes due to treatment.

protein_abundance_normalisation(
  lip_peptides,
  tc_proteins,
  sample,
  grouping,
  condition,
  protein_id,
  peptide_intensity,
  protein_intensity,
  method,
  paired = FALSE,
  replicate_index = NULL,
  comparison = NULL,
  completeness = 0.7
)

Arguments

lip_peptides
A data frame containing LiP peptide level data. The sample, protein_id, grouping, condition, peptide_intensity variables are required. Ideally the output of assign_missingness or impute is used for the method "ref_vs_rest", since here also the variable missingness is required in contrast to method "all_vs_all".

tc_proteins
A data frame containing tryptic control protein level data. The sample, protein_id, condition, protein_intensity variables are required. Ideally the output of assign_missingness or impute is used for the method "ref_vs_rest", since here also the variable missingness is required in contrast to method "all_vs_all".

sample
The column in the data frame containing the sample name. Is not required if method = "t-test_mean_sd".

grouping
The column in the data frame containing precursor or peptide identifiers.

condition
The column in the data frame containing the conditions.

protein_id
The column in the data frame containing the protein accession numbers.

peptide_intensity
The column in the data frame containing peptide intensity values. The intensity values need to be log2 transformed.

protein_intensity
The column in the data frame containing protein intensity values. The intensity values need to be log2 transformed.

method
A character vector specifying which method should be used. If method = "all_vs_all" an ANOVA based method will be performed that compares all conditions with each other followed by Tukey's honestly significant difference test to identify pairs of conditions that are significantly different. If method = "ref_vs_rest" a reference condition will be compared against all other conditions using a Welch t-test in order to identify significantly changing peptides or precursors.

paired
A logical indicating if LiP and tryptic control replicates for each condition are paired. That means they belong together because they were taken from the exact same sample and then split in LiP and tryptic control conditions. If they are paired protein abundances are only used if there was a corresponding peptide detected in the LiP condition.

replicate_index
The column in the data frame containing numbering for condition replicates. This column is only required if LiP and tryptic control samples are paired. In that case the corresponding samples in both files should have the same numbering.

comparison
The column in the data frame containing comparison information of treatment/reference condition pairs. Can be obtained by calling assign_missingness.

completeness
The minimal degree of data completeness to be considered for evaluation. Value has to be between 0 and 1, default is 0.7. It is multiplied with the number of replicates and then adjusted downward. The resulting number is the minimal number of observations for each condition to be considered for evaluation.

Value

A data frame that contains significance test statistics and differential abundances (diff) for each peptide or precursor and the corresponding comparison of conditions. If method = "all_vs_all" was used, ANOVA statistics are reported (ms_group, ms_error, f, pval, adj_pval, n_conditions). ms_group is the error within a group and ms_error is the error between groups. Also Tukey's honestly significant difference test statistics are reported (df, q_value, pval_tukey). If method = "ref_vs_rest" Welch t-test statistics are reported (std_error, t_statistic, pval, adj_pval).

Examples

if (FALSE) {
protein_abundance_normalisation(
lip_peptides = data_lip, 
tc_proteins = data_tc, 
sample = r_file_name,
protein_id = pg_protein_accessions, 
grouping = eg_precursor_id, 
condition = r_condition,
peptide_intensity = normalised_intensity_log2,
protein_intensity = protein_intensity_log2,
method = "all_vs_all")
}