Check CV distribution

Calculates and plots the coefficients of variation for the selected grouping.

qc_cvs(
  data,
  grouping,
  condition,
  intensity,
  plot = TRUE,
  plot_style = "density"
)

Arguments

data
A dataframe containing at least peptide, precursor or protein identifiers, information on conditions and intensity values for each peptide, precursor or protein.

grouping
The column in the input dataframe containing the grouping variables (e.g. peptides, precursors or proteins).

condition
The column in the data dataframe containing condition information (e.g. "treated" and "control").

intensity
Column containing the corresponding raw intensity values (not log2 transformed) for each peptide or precursor.

plot
A logical indicating whether the result should be plotted. Default is TRUE.

plot_style
Variable indicating the plotting style. plot_style = "boxplot" plots a boxplot, whereas plot_style = "density" plots the CV density distribution. plot_style = "violin" returns a violin plot. Default is plot_style = "density".

Value

Either the median CVs in

Examples

if (FALSE) {
qc_cvs(
data,
grouping = pep_stripped_sequence,
condition = condition,
intensity = fg_quantity,
plot = TRUE,
plot_style = "density"
)
}