Correlation based hirachical clustering of samples

A correlation heatmap is created that uses hirachical clustering to determine sample similarity.

qc_sample_correlation(
  data,
  sample,
  grouping,
  intensity,
  condition,
  digestion = NULL,
  run_order = NULL,
  method = "spearman",
  interactive = FALSE
)

Arguments

data
a dataframe contains at least the input variables.

sample
the name of the column containing the sample names.

grouping
the name of the column containing precursor or peptide identifiers.

intensity
the name of the column containing intensity values. Note: The input intensities should be log2 transformed.

condition
the name of the column containing the conditions.

digestion
optional, the name of the column containing information about the digestion method used. Eg. "LiP" or "tryptic control".

run_order
optional, the name of the column containing the order in which samples were measured. Useful to investigate batch effects due to run order.

method
the method to be used for correlation. "spearman" is the default but can be changed to "pearson" or "kendall".

interactive
logical, default is FALSE. Determines if an interactive or static heatmap should be created using heatmaply or pheatmap, respectively.

Value

A correlation heatmap that compares each sample. The dendrogram is sorted by optimal leaf ordering.

Examples

if (FALSE) {
qc_sample_correlation(
data,
sample = r_file_name,
grouping = eg_precursor_id,
intensity = intensity_log2,
condition = r_condition
)
}