Perform KEGG pathway enrichment analysis

Analyses enrichment of KEGG pathways associated with proteins in the fraction of significant proteins compared to all detected proteins. A Fisher's exact test is performed to test significance of enrichment.

calculate_kegg_enrichment(
  data,
  protein_id,
  is_significant,
  pathway_id = pathway_id,
  pathway_name = pathway_name,
  plot = TRUE,
  plot_cutoff = "adj_pval top10"
)

Arguments

data

a data frame that contains at least the input variables.

protein_id

a character column in the data data frame that contains the protein accession numbers.

is_significant

a logical column in the data data frame that indicates if the corresponding protein has a significantly changing peptide. The input data frame may contain peptide level information with significance information. The function is able to extract protein level information from this.

pathway_id

a character column in the data data frame that contains KEGG pathway identifiers. These can be obtained from KEGG using fetch_kegg.

pathway_name

a character column in the data data frame that contains KEGG pathway names. These can be obtained from KEGG using fetch_kegg.

plot

a logical value indicating whether the result should be plotted or returned as a table.

plot_cutoff

a character value indicating if the plot should contain the top 10 most significant proteins (p-value or adjusted p-value), or if a significance cutoff should be used to determine the number of GO terms in the plot. This information should be provided with the type first followed by the threshold separated by a space. Example are plot_cutoff = "adj_pval top10", plot_cutoff = "pval 0.05" or plot_cutoff = "adj_pval 0.01". The threshold can be chosen freely.

Value

A bar plot displaying negative log10 adjusted p-values for the top 10 enriched pathways. Bars are coloured according to the direction of the enrichment. If plot = FALSE, a data frame is returned.

Examples

# \donttest{
# Load libraries
library(dplyr)

set.seed(123) # Makes example reproducible

# Create example data
kegg_data <- fetch_kegg(species = "eco")

if (!is.null(kegg_data)) { # only proceed if information was retrieved
  data <- kegg_data %>%
    group_by(uniprot_id) %>%
    mutate(significant = rep(sample(
      x = c(TRUE, FALSE),
      size = 1,
      replace = TRUE,
      prob = c(0.2, 0.8)
    ),
    n = n()
    ))

  # Plot KEGG enrichment
  calculate_kegg_enrichment(
    data,
    protein_id = uniprot_id,
    is_significant = significant,
    pathway_id = pathway_id,
    pathway_name = pathway_name,
    plot = TRUE,
    plot_cutoff = "pval 0.05"
  )

  # Calculate KEGG enrichment
  kegg <- calculate_kegg_enrichment(
    data,
    protein_id = uniprot_id,
    is_significant = significant,
    pathway_id = pathway_id,
    pathway_name = pathway_name,
    plot = FALSE
  )

  head(kegg, n = 10)
}
#> # A tibble: 10 × 10
#>    pathway_id    pathw…¹    pval adj_p…² n_det…³ n_det…⁴ n_sig…⁵ n_sig…⁶ n_pro…⁷
#>    <chr>         <chr>     <dbl>   <dbl>   <int>   <int>   <int>   <int>   <dbl>
#>  1 path:eco01503 Cation… 0.00926       1    1615      36     314      14    7   
#>  2 path:eco00230 Purine… 0.0197        1    1615      78     314      24   15.2 
#>  3 path:eco01232 Nucleo… 0.0515        1    1615      63     314      19   12.2 
#>  4 path:eco00410 beta-A… 0.0858        1    1615      14     314       0    2.72
#>  5 path:eco00670 One ca… 0.0859        1    1615      13     314       0    2.53
#>  6 path:eco00592 alpha-… 0.0994        1    1615       3     314       2    0.58
#>  7 path:eco00770 Pantot… 0.104         1    1615      23     314       1    4.47
#>  8 path:eco00750 Vitami… 0.113         1    1615      10     314       4    1.94
#>  9 path:eco02030 Bacter… 0.151         1    1615      20     314       1    3.89
#> 10 path:eco02060 Phosph… 0.158         1    1615      41     314       4    7.97
#> # … with 1 more variable: direction <chr>, and abbreviated variable names
#> #   ¹​pathway_name, ²​adj_pval, ³​n_detected_proteins,
#> #   ⁴​n_detected_proteins_in_pathway, ⁵​n_significant_proteins,
#> #   ⁶​n_significant_proteins_in_pathway, ⁷​n_proteins_expected
# }