Fetch AlphaFold prediction

Fetches atom level data for AlphaFold predictions either for selected proteins or whole organisms.

fetch_alphafold_prediction(
  uniprot_ids = NULL,
  organism_name = NULL,
  version = "v4",
  timeout = 3600,
  return_data_frame = FALSE,
  show_progress = TRUE
)

Arguments

uniprot_ids

optional, a character vector of UniProt identifiers for which predictions should be fetched. This argument is mutually exclusive to the organism_name argument.

organism_name

optional, a character value providing the name of an organism for which all available AlphaFold predictions should be retreived. The name should be the capitalised scientific species name (e.g. "Homo sapiens"). Note: Some organisms contain a lot of predictions which might take a considerable amount of time and memory to fetch. Therefore, you should be sure that your system can handle fetching predictions for these organisms. This argument is mutually exclusive to the uniprot_ids argument.

version

a character value that specifies the alphafold version that should be used. This is regularly updated by the database. We always try to make the current version the default version. Available version can be found here: https://ftp.ebi.ac.uk/pub/databases/alphafold/

timeout

a numeric value specifying the time in seconds until the download of an organism archive times out. The default is 3600 seconds.

return_data_frame

a logical value that specifies if true, a data frame instead of a list is returned. It is recommended to only use this if information for few proteins is retrieved. Default is FALSE.

show_progress

a logical value that specifies if true, a progress bar will be shown. Default is TRUE.

Value

A list that contains atom level data for AlphaFold predictions. If return_data_frame is TRUE, a data frame with this information is returned instead. The data frame contains the following columns:

label_id: Uniquely identifies every atom in the prediction following the standardised convention for mmCIF files.

type_symbol: The code used to identify the atom species representing this atom type. This code is the element symbol.

label_atom_id: Uniquely identifies every atom for the given residue following the standardised convention for mmCIF files.

label_comp_id: A chemical identifier for the residue. This is the three- letter code for the amino acid.

label_asym_id: Chain identifier following the standardised convention for mmCIF files. Since every prediction only contains one protein this is always "A".

label_seq_id: Uniquely and sequentially identifies residues for each protein. The numbering corresponds to the UniProt amino acid positions.

x: The x coordinate of the atom.

y: The y coordinate of the atom.

z: The z coordinate of the atom.

prediction_score: Contains the prediction score for each residue.

auth_seq_id: Same as label_seq_id. But of type character.

auth_comp_id: Same as label_comp_id.

auth_asym_id: Same as label_asym_id.

uniprot_id: The UniProt identifier of the predicted protein.

score_quality: Score annotations.

Examples

# \donttest{
alphafold <- fetch_alphafold_prediction(
  uniprot_ids = c("F4HVG8", "O15552"),
  return_data_frame = TRUE
)

head(alphafold, n = 10)
#> # A tibble: 10 × 15
#>    label_id type_sym…¹ label…² label…³ label…⁴ label…⁵     x     y     z predi…⁶
#>       <dbl> <chr>      <chr>   <chr>   <chr>     <dbl> <dbl> <dbl> <dbl>   <dbl>
#>  1        1 N          N       MET     A             1 -15.4  18.3 -11.7    27.3
#>  2        2 C          CA      MET     A             1 -15.8  17.4 -12.8    27.3
#>  3        3 C          C       MET     A             1 -15.5  16.0 -12.5    27.3
#>  4        4 C          CB      MET     A             1 -15.1  17.8 -14.1    27.3
#>  5        5 O          O       MET     A             1 -14.4  15.8 -12.0    27.3
#>  6        6 C          CG      MET     A             1 -15.2  19.2 -14.6    27.3
#>  7        7 S          SD      MET     A             1 -14.3  19.5 -16.2    27.3
#>  8        8 C          CE      MET     A             1 -14.3  21.3 -16.3    27.3
#>  9        9 N          N       LEU     A             2 -16.3  15.0 -12.9    26.2
#> 10       10 C          CA      LEU     A             2 -17.7  14.9 -13.2    26.2
#> # … with 5 more variables: auth_seq_id <chr>, auth_comp_id <chr>,
#> #   auth_asym_id <chr>, uniprot_id <chr>, score_quality <chr>, and abbreviated
#> #   variable names ¹​type_symbol, ²​label_atom_id, ³​label_comp_id,
#> #   ⁴​label_asym_id, ⁵​label_seq_id, ⁶​prediction_score
# }