Fetch protein disorder and mobility information from MobiDB

Fetches information about disordered and flexible protein regions from MobiDB.

fetch_mobidb(uniprot_ids = NULL, organism_id = NULL, show_progress = TRUE)

Arguments

uniprot_ids

optional, a character vector of UniProt identifiers for which information should be fetched. This argument is mutually exclusive to the organism_id argument.

organism_id

optional, a character value providing the NCBI taxonomy identifier of an organism (TaxId) of an organism for which all available information should be retreived. This argument is mutually exclusive to the uniprot_ids argument.

show_progress

a logical value; if TRUE a progress bar will be shown. Default is TRUE.

Value

A data frame that contains start and end positions for disordered and flexible protein regions. The feature column contains information on the source of this annotation. More information on the source can be found here.

Examples

# \donttest{
fetch_mobidb(
  uniprot_ids = c("P0A799", "P62707")
)
#> # A tibble: 95 × 6
#>    accession feature                  start end   content_fraction content_count
#>    <chr>     <chr>                    <chr> <chr>            <dbl>         <dbl>
#>  1 P0A799    derived-observed-th_90   2     387              0.997           386
#>  2 P0A799    prediction-plddt-alphaf… 2     387              0.997           386
#>  3 P0A799    homology-domain-gene3d   165   373              0.966           374
#>  4 P0A799    homology-domain-gene3d   8     381              0.966           374
#>  5 P0A799    homology-domain-merge    6     381              0.972           376
#>  6 P0A799    homology-domain-pfam     6     373              0.951           368
#>  7 P0A799    prediction-disorder-dis… 1     5                0.052            20
#>  8 P0A799    prediction-disorder-dis… 118   123              0.052            20
#>  9 P0A799    prediction-disorder-dis… 146   150              0.052            20
#> 10 P0A799    prediction-disorder-dis… 384   387              0.052            20
#> # ℹ 85 more rows
# }