Fetch protein data from UniProt

Fetches protein metadata from UniProt.

fetch_uniprot(
  uniprot_ids,
  columns = c("protein_name", "length", "sequence", "gene_names", "xref_geneid",
    "xref_string", "go_f", "go_p", "go_c", "cc_interaction", "ft_act_site", "ft_binding",
    "cc_cofactor", "cc_catalytic_activity", "xref_pdb"),
  batchsize = 200,
  show_progress = TRUE
)

Arguments

uniprot_ids

a character vector of UniProt accession numbers.

columns

a character vector of metadata columns that should be imported from UniProt (all possible columns can be found here. For cross-referenced database provide the database name with the prefix "xref_", e.g. "xref_pdb")

batchsize

a numeric value that specifies the number of proteins processed in a single single query. Default and max value is 200.

show_progress

a logical value that determines if a progress bar will be shown. Default is TRUE.

Value

A data frame that contains all protein metadata specified in columns for the proteins provided. If an invalid ID was provided that contains a valid UniProt ID, the valid portion of the ID is fetched and the invalid input ID is saved in a column called input_id.

Examples

# \donttest{
fetch_uniprot(c("P36578", "O43324", "Q00796"))
#> Please note that some column names have changed due to UniProt updating its API! This might cause errors in your code. You can fix it by replacing the old column names with new ones.
#> # A tibble: 3 × 16
#>   accession protein_n…¹ length seque…² gene_…³ xref_…⁴ xref_…⁵ go_f  go_p  go_c 
#>   <chr>     <chr>        <dbl> <chr>   <chr>   <chr>   <chr>   <chr> <chr> <chr>
#> 1 O43324    Eukaryotic…    174 MAAAAE… EEF1E1… 9521;   9606.E… NA    cell… amin…
#> 2 P36578    60S riboso…    427 MACARP… RPL4 R… 6124;   9606.E… RNA … cyto… cyto…
#> 3 Q00796    Sorbitol d…    357 MAAAAK… SORD    6652;   9606.E… (R,R… flag… cyto…
#> # … with 6 more variables: cc_interaction <chr>, ft_act_site <lgl>,
#> #   ft_binding <chr>, cc_cofactor <chr>, cc_catalytic_activity <chr>,
#> #   xref_pdb <chr>, and abbreviated variable names ¹​protein_name, ²​sequence,
#> #   ³​gene_names, ⁴​xref_geneid, ⁵​xref_string
# }