Imputation of missing values

impute is calculating imputation values for missing data depending on the selected method.

impute(
  data,
  sample,
  grouping,
  intensity_log2,
  condition,
  comparison = comparison,
  missingness = missingness,
  noise = NULL,
  method = "ludovic",
  skip_log2_transform_error = FALSE,
  retain_columns = NULL
)

Arguments

data

a data frame that is ideally the output from the assign_missingness function. It should containing at least the input variables. For each "reference_vs_treatment" comparison, there should be the pair of the reference and treatment condition. That means the reference condition should be doublicated once for every treatment.

sample

a character column in the data data frame that contains the sample names.

grouping

a character column in the data data frame that contains the precursor or peptide identifiers.

intensity_log2

a numeric column in the data data frame that contains the intensity values.

condition

a character or numeric column in the data data frame that contains the the conditions.

comparison

a character column in the data data frame that contains the the comparisons of treatment/reference pairs. This is an output of the assign_missingnes function.

missingness

a character column in the data data frame that contains the missingness type of the data determines how values for imputation are sampled. This should at least contain "MAR" or "MNAR". Missingness assigned as NA will not be imputed.

noise

a numeric column in the data data frame that contains the noise value for the precursor/peptide. Is only required if method = "noise". Note: Noise values need to be log2 transformed.

method

a character value that specifies the method to be used for imputation. For method = "ludovic", MNAR missingness is sampled from a normal distribution around a value that is three lower (log2) than the lowest intensity value recorded for the precursor/peptide and that has a spread of the mean standard deviation for the precursor/peptide. For method = "noise", MNAR missingness is sampled from a normal distribution around the mean noise for the precursor/peptide and that has a spread of the mean standard deviation (from each condition) for the precursor/peptide. Both methods impute MAR data using the mean and variance of the condition with the missing data.

skip_log2_transform_error

a logical value that determines if a check is performed to validate that input values are log2 transformed. If input values are > 40 the test is failed and an error is returned.

retain_columns

a vector that indicates columns that should be retained from the input data frame. Default is not retaining additional columns retain_columns = NULL. Specific columns can be retained by providing their names (not in quotations marks, just like other column names, but in a vector).

Value

A data frame that contains an imputed_intensity and imputed column in addition to the required input columns. The imputed column indicates if a value was imputed. The imputed_intensity column contains imputed intensity values for previously missing intensities.

Examples

set.seed(123) # Makes example reproducible

# Create example data
data <- create_synthetic_data(
  n_proteins = 10,
  frac_change = 0.5,
  n_replicates = 4,
  n_conditions = 2,
  method = "effect_random",
  additional_metadata = FALSE
)

head(data, n = 24)
#> # A tibble: 24 × 8
#>    protein   peptide     condition   sample   peptide_i…¹ change chang…² pepti…³
#>    <chr>     <chr>       <chr>       <chr>          <dbl> <lgl>  <lgl>     <dbl>
#>  1 protein_1 peptide_1_1 condition_1 sample_1        16.8 TRUE   TRUE       16.8
#>  2 protein_1 peptide_1_1 condition_1 sample_2        17.0 TRUE   TRUE       17.0
#>  3 protein_1 peptide_1_1 condition_1 sample_3        17.0 TRUE   TRUE       17.0
#>  4 protein_1 peptide_1_1 condition_1 sample_4        17.0 TRUE   TRUE       17.0
#>  5 protein_1 peptide_1_1 condition_2 sample_5        15.8 TRUE   TRUE       15.8
#>  6 protein_1 peptide_1_1 condition_2 sample_6        15.9 TRUE   TRUE       15.9
#>  7 protein_1 peptide_1_1 condition_2 sample_7        16.1 TRUE   TRUE       16.1
#>  8 protein_1 peptide_1_1 condition_2 sample_8        15.9 TRUE   TRUE       15.9
#>  9 protein_1 peptide_1_2 condition_1 sample_1        12.6 TRUE   FALSE      NA  
#> 10 protein_1 peptide_1_2 condition_1 sample_2        12.7 TRUE   FALSE      NA  
#> # … with 14 more rows, and abbreviated variable names ¹​peptide_intensity,
#> #   ²​change_peptide, ³​peptide_intensity_missing

# Assign missingness information
data_missing <- assign_missingness(
  data,
  sample = sample,
  condition = condition,
  grouping = peptide,
  intensity = peptide_intensity_missing,
  ref_condition = "all",
  retain_columns = c(protein, peptide_intensity)
)
#> "all" was provided as reference condition. All pairwise comparisons are
#> created from the conditions and assigned their missingness. The created
#> comparisons are:
#> condition_1_vs_condition_2

head(data_missing, n = 24)
#> # A tibble: 24 × 8
#>    protein   peptide_intensity sample   condit…¹ peptide pepti…² compa…³ missi…⁴
#>    <chr>                 <dbl> <chr>    <chr>    <chr>     <dbl> <chr>   <chr>  
#>  1 protein_1              16.8 sample_1 conditi… peptid…    16.8 condit… comple…
#>  2 protein_1              17.0 sample_2 conditi… peptid…    17.0 condit… comple…
#>  3 protein_1              17.0 sample_3 conditi… peptid…    17.0 condit… comple…
#>  4 protein_1              17.0 sample_4 conditi… peptid…    17.0 condit… comple…
#>  5 protein_1              15.8 sample_5 conditi… peptid…    15.8 condit… comple…
#>  6 protein_1              15.9 sample_6 conditi… peptid…    15.9 condit… comple…
#>  7 protein_1              16.1 sample_7 conditi… peptid…    16.1 condit… comple…
#>  8 protein_1              15.9 sample_8 conditi… peptid…    15.9 condit… comple…
#>  9 protein_1              12.6 sample_1 conditi… peptid…    NA   condit… NA     
#> 10 protein_1              12.7 sample_2 conditi… peptid…    NA   condit… NA     
#> # … with 14 more rows, and abbreviated variable names ¹​condition,
#> #   ²​peptide_intensity_missing, ³​comparison, ⁴​missingness

# Perform imputation
data_imputed <- impute(
  data_missing,
  sample = sample,
  grouping = peptide,
  intensity_log2 = peptide_intensity_missing,
  condition = condition,
  comparison = comparison,
  missingness = missingness,
  method = "ludovic",
  retain_columns = c(protein, peptide_intensity)
)

head(data_imputed, n = 24)
#> # A tibble: 24 × 10
#>    protein   peptide_in…¹ sample condi…² peptide pepti…³ compa…⁴ missi…⁵ imput…⁶
#>    <chr>            <dbl> <chr>  <chr>   <chr>     <dbl> <chr>   <chr>     <dbl>
#>  1 protein_1         16.8 sampl… condit… peptid…    16.8 condit… comple…    16.8
#>  2 protein_1         17.0 sampl… condit… peptid…    17.0 condit… comple…    17.0
#>  3 protein_1         17.0 sampl… condit… peptid…    17.0 condit… comple…    17.0
#>  4 protein_1         17.0 sampl… condit… peptid…    17.0 condit… comple…    17.0
#>  5 protein_1         15.8 sampl… condit… peptid…    15.8 condit… comple…    15.8
#>  6 protein_1         15.9 sampl… condit… peptid…    15.9 condit… comple…    15.9
#>  7 protein_1         16.1 sampl… condit… peptid…    16.1 condit… comple…    16.1
#>  8 protein_1         15.9 sampl… condit… peptid…    15.9 condit… comple…    15.9
#>  9 protein_1         12.6 sampl… condit… peptid…    NA   condit… NA         NA  
#> 10 protein_1         12.7 sampl… condit… peptid…    NA   condit… NA         NA  
#> # … with 14 more rows, 1 more variable: imputed <lgl>, and abbreviated variable
#> #   names ¹​peptide_intensity, ²​condition, ³​peptide_intensity_missing,
#> #   ⁴​comparison, ⁵​missingness, ⁶​imputed_intensity