Randomise samples in MS queue

This function randomises the order of samples in an MS queue. QC and Blank samples are left in place. It is also possible to randomise only parts of the queue. Before running this make sure to set a specific seed with the set.seed() function. This ensures that the randomisation of the result is consistent if the function is run again.

randomise_queue(data = NULL, rows = NULL, export = FALSE)

Arguments

data

optional, a data frame that contains a queue. If not provided a queue file can be chosen interactively.

rows

optional, a numeric vector that specifies a range of rows in for which samples should be randomized.

export

a logical value that determines if a "randomised_queue.csv" file will be saved in the working directory. If FALSE a data frame will be returned.

Value

If export = TRUE a "randomised_queue.csv" file will be saved in the working directory. If export = FALSE a data frame that contains the randomised queue is returned.

Examples

queue <- create_queue(
  date = c("200722"),
  instrument = c("EX1"),
  user = c("jquast"),
  measurement_type = c("DIA"),
  experiment_name = c("JPQ031"),
  digestion = c("LiP", "tryptic control"),
  treatment_type_1 = c("EDTA", "H2O"),
  treatment_type_2 = c("Zeba", "unfiltered"),
  treatment_dose_1 = c(10, 30, 60),
  treatment_unit_1 = c("min"),
  n_replicates = 4,
  number_runs = FALSE,
  organism = c("E. coli"),
  exclude_combinations = list(list(
    treatment_type_1 = c("H2O"),
    treatment_type_2 = c("Zeba", "unfiltered"),
    treatment_dose_1 = c(10, 30)
  )),
  inj_vol = c(2),
  data_path = "D:\\2007_Data",
  method_path = "C:\\Xcalibur\\methods\\DIA_120min",
  position_row = c("A", "B", "C", "D", "E", "F"),
  position_column = 8,
  blank_every_n = 4,
  blank_position = "1-V1",
  blank_method_path = "C:\\Xcalibur\\methods\\blank"
)

head(queue, n = 20)
#> # A tibble: 20 × 21
#>    Sample …¹ File …² Sampl…³ Path  Instr…⁴ Proce…⁵ Calib…⁶ Posit…⁷ Inj V…⁸ Level
#>    <chr>     <chr>     <dbl> <chr> <chr>   <lgl>   <lgl>   <chr>     <dbl> <lgl>
#>  1 QC        200722…       1 "D:\… "C:\\X… NA      NA      1-V1          1 NA   
#>  2 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      A1            2 NA   
#>  3 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      A2            2 NA   
#>  4 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      A3            2 NA   
#>  5 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      A4            2 NA   
#>  6 QC        200722…       1 "D:\… "C:\\X… NA      NA      1-V1          1 NA   
#>  7 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      A5            2 NA   
#>  8 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      A6            2 NA   
#>  9 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      A7            2 NA   
#> 10 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      A8            2 NA   
#> 11 QC        200722…       1 "D:\… "C:\\X… NA      NA      1-V1          1 NA   
#> 12 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      B1            2 NA   
#> 13 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      B2            2 NA   
#> 14 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      B3            2 NA   
#> 15 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      B4            2 NA   
#> 16 QC        200722…       1 "D:\… "C:\\X… NA      NA      1-V1          1 NA   
#> 17 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      B5            2 NA   
#> 18 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      B6            2 NA   
#> 19 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      B7            2 NA   
#> 20 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      B8            2 NA   
#> # … with 11 more variables: `Sample Wt` <dbl>, `Sample Vol` <dbl>,
#> #   `ISTD Amt` <dbl>, `Dil Factor` <dbl>, `L1 Study` <lgl>, `L2 Client` <lgl>,
#> #   `L3 Laboratory` <lgl>, `L4 Company` <lgl>, `L5 Phone` <lgl>, Comment <lgl>,
#> #   `Sample Name` <chr>, and abbreviated variable names ¹​`Sample Type`,
#> #   ²​`File Name`, ³​`Sample ID`, ⁴​`Instrument Method`, ⁵​`Process Method`,
#> #   ⁶​`Calibration File`, ⁷​Position, ⁸​`Inj Vol`

randomised_queue <- randomise_queue(
  data = queue,
  export = FALSE
)

head(randomised_queue, n = 20)
#> # A tibble: 20 × 21
#>    Sample …¹ File …² Sampl…³ Path  Instr…⁴ Proce…⁵ Calib…⁶ Posit…⁷ Inj V…⁸ Level
#>    <chr>     <chr>     <dbl> <chr> <chr>   <lgl>   <lgl>   <chr>     <dbl> <lgl>
#>  1 QC        200722…       1 "D:\… "C:\\X… NA      NA      1-V1          1 NA   
#>  2 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      C8            2 NA   
#>  3 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      F6            2 NA   
#>  4 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      A4            2 NA   
#>  5 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      B2            2 NA   
#>  6 QC        200722…       1 "D:\… "C:\\X… NA      NA      1-V1          1 NA   
#>  7 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      F5            2 NA   
#>  8 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      E6            2 NA   
#>  9 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      D4            2 NA   
#> 10 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      A6            2 NA   
#> 11 QC        200722…       1 "D:\… "C:\\X… NA      NA      1-V1          1 NA   
#> 12 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      A5            2 NA   
#> 13 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      F4            2 NA   
#> 14 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      A2            2 NA   
#> 15 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      D3            2 NA   
#> 16 QC        200722…       1 "D:\… "C:\\X… NA      NA      1-V1          1 NA   
#> 17 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      C5            2 NA   
#> 18 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      F8            2 NA   
#> 19 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      B5            2 NA   
#> 20 Unknown   200722…       1 "D:\… "C:\\X… NA      NA      B4            2 NA   
#> # … with 11 more variables: `Sample Wt` <dbl>, `Sample Vol` <dbl>,
#> #   `ISTD Amt` <dbl>, `Dil Factor` <dbl>, `L1 Study` <lgl>, `L2 Client` <lgl>,
#> #   `L3 Laboratory` <lgl>, `L4 Company` <lgl>, `L5 Phone` <lgl>, Comment <lgl>,
#> #   `Sample Name` <chr>, and abbreviated variable names ¹​`Sample Type`,
#> #   ²​`File Name`, ³​`Sample ID`, ⁴​`Instrument Method`, ⁵​`Process Method`,
#> #   ⁶​`Calibration File`, ⁷​Position, ⁸​`Inj Vol`